ocr/requirements.txt
# Dependencies for the OCR container.

fastapi==0.115.12
uvicorn[standard]==0.34.2
python-doctr==0.11.0
pillow==11.2.1
beautifulsoup4==4.13.4
anyascii==0.3.2
certifi==2025.4.26
charset-normalizer==3.4.2
defusedxml==0.7.1
filelock==3.18.0
fsspec==2025.5.0
h5py==3.13.0
huggingface-hub==0.31.4
idna==3.10
Jinja2==3.1.6
langdetect==1.0.9
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy<2.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnx==1.18.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
protobuf==6.31.0
pyclipper==1.3.0.post6
pypdfium2==4.30.1
python-doctr==0.11.0
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
scipy==1.15.3
setuptools==80.8.0
shapely==2.1.1
six==1.17.0
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing_extensions==4.13.2
urllib3==2.4.0
wheel==0.45.1